src/security-advisor/utils.py
import os
from dotenv import load_dotenv

def load_github_token():
    load_dotenv()
    token = os.getenv('GITHUB_TOKEN')
    if not token:
        raise ValueError("GitHub token not found in environment variables")
    return token
